+++ plugins/PepperFlashPlayer-mac.plugin/Contents/Info.plist
