+++ plugins/PepperFlashPlayer-mac.plugin/Contents/_CodeSignature/CodeResources
