+++ windows/about/Spotify_Logo_RGB_White.png
