+++ windows/spotify/D-Bus/interpreter.js
