+++ windows/spotify/Sing!/sing.js
