+++ windows/spotify/base-theme.css
