+++ windows/spotify/controller.js
