+++ windows/spotify/interface.js
