+++ windows/spotify/jquery.js
