+++ windows/spotify/preload.js
