+++ windows/spotify/themes/dark-theme.css
