+++ windows/spotify/themes/light-theme.css
