+++ windows/spotify/tray.js
