+++ windows/spotify/user.js
