+++ windows/spotify/window-menu.js
